Consolidate team modal form into component
